カスの修正

+++ src/main/java/model/dto/SalesDTO.java
@@ -2,46 +2,37 @@ package model.dto;
 
 import java.io.Serializable;
 
-/**
- * 売上データを保持するDTOクラス
- */
 public class SalesDTO implements Serializable {
 	private static final long serialVersionUID = 1L;
 
 	private String displayDate;
-	private int totalAmount;
-	private int visitorCount;
-	private int dayOfWeek; // 1:月 ... 6:土, 7:日
+	private long totalAmount; // ★ int から long へ変更
+	private long visitorCount; // ★ 念のため来店数も long へ変更
+	private int dayOfWeek;
 
-	// 基本コンストラクタ
-	public SalesDTO(String displayDate, int totalAmount, int visitorCount) {
+	public SalesDTO(String displayDate, long totalAmount, long visitorCount) {
 		this.displayDate = displayDate;
 		this.totalAmount = totalAmount;
 		this.visitorCount = visitorCount;
 	}
 
-	// 曜日情報付きコンストラクタ（日間表示で使用）
-	public SalesDTO(String displayDate, int totalAmount, int visitorCount, int dayOfWeek) {
+	public SalesDTO(String displayDate, long totalAmount, long visitorCount, int dayOfWeek) {
 		this(displayDate, totalAmount, visitorCount);
 		this.dayOfWeek = dayOfWeek;
 	}
 
-	// JSPからプロパティとしてアクセスするためのゲッター
 	public String getDisplayDate() {
 		return displayDate;
 	}
 
-	public int getTotalAmount() {
+	public long getTotalAmount() {
 		return totalAmount;
-	}
+	} // ★ long
 
-	public int getVisitorCount() {
+	public long getVisitorCount() {
 		return visitorCount;
-	}
+	} // ★ long
 
-	/**
-	 * JSPの ${s.dayOfWeek} で呼び出されるメソッド
-	 */
 	public int getDayOfWeek() {
 		return dayOfWeek;
 	}
@@ -49,7 +40,7 @@ public class SalesDTO implements Serializable {
 	/**
 	 * 客単価を計算するメソッド
 	 */
-	public int getAveragePrice() {
+	public long getAveragePrice() { // ★ long
 		if (visitorCount == 0)
 			return 0;
 		return totalAmount / visitorCount;
